Droplet+Setup: move route setup to this file
